Extract query argument parser partial
